requirements.txt
pillow>=11.3.0
psycopg2-binary>=2.9.9
requests>=2.32.5
tweepy>=4.16.0
